+++ R/lmtp_control.R
@@ -11,8 +11,6 @@
 #'  The number of cross-validation folds for \code{learners_outcome}.
 #' @param .learners_trt_folds \[\code{integer(1)}\]\cr
 #'  The number of cross-validation folds for \code{learners_trt}.
-#' @param .return_full_fits \[\code{logical(1)}\]\cr
-#'  Return full 'SuperLearner' fits? Default is \code{FALSE}, return only 'SuperLearner' coefficients.
 #' @param .discrete \[\code{logical(1)}\]\cr
 #'  Use discrete or ensemble super learner?
 #' @param .info \[\code{logical(1)}\]\cr
@@ -27,7 +25,6 @@ lmtp_control <- function(.bound = 1e5,
                          .trim = 0.999,
                          .learners_outcome_folds = 10,
                          .learners_trt_folds = 10,
-                         .return_full_fits = FALSE,
                          .discrete = TRUE,
                          .info = FALSE) {
 
@@ -35,7 +32,6 @@ lmtp_control <- function(.bound = 1e5,
   assert_number(.learners_trt_folds, null.ok = TRUE)
   assert_number(.bound)
   assert_number(.trim, upper = 1)
-  assert_logical(.return_full_fits, len = 1)
   assert_logical(.discrete, len = 1)
   assert_logical(.info, len = 1)
 
@@ -43,7 +39,6 @@ lmtp_control <- function(.bound = 1e5,
        .trim = .trim,
        .learners_outcome_folds = .learners_outcome_folds,
        .learners_trt_folds = .learners_trt_folds,
-       .return_full_fits = .return_full_fits,
        .discrete = .discrete,
        .info = .info)
 }